+++ src/pages/LogPage/LogPage.module.css
@@ -1,117 +1,4 @@
-.topwrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  justify-content: space-between;
-  align-content: center;
-  margin-bottom: 20px;
-}
-
-.top {
-  display: flex;
-  justify-content: space-between;
-}
-
-.linkContainer {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.dateWrapper {
-  display: flex;
-  justify-content: space-between;
-}
-
-.timeContainer {
-  display: flex;
-  flex-direction: column;
-  font-size: 18px;
-  color: var(--gray_818181);
-}
-
-.totalTime {
-  font-size: 28px;
-  font-weight: 700;
-  margin-top: 4px;
-}
-
-/* 집중 시간, 포인트 */
-.radioBox {
-  display: flex;
-  justify-content: end;
-  align-items: end;
-  appearance: none;
-  gap: 16px;
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.radioBox input { 
-  display: none;
-}
-
-.radioBoxItem {
-  display: flex;
-  cursor: pointer;
-  align-items: center;
-  padding: 6px 18px;
-  border-radius: 30px;
-  border: 1px solid var(--gray_EEEEEE);
-  background: var(--gray_EEEEEE);
-  color: var(--gray_818181);
-}
-
-.active {
-  background: var(--brand_99C08E);
-  border: 1px solid var(--brand_99C08E);
-  color: var(--white_FFFFFF);
-  border: none;
-}
-
-/* 날짜 이동 컨테이너 */
-.dateSelectorContainer {
-  display: flex;
-  justify-content: space-between;
-}
-
-.nowDate {
-  display: flex;
-  align-items: center;
-  font-size: 20px;
-  font-weight: 600;
-  color: var(--gray_818181);
-}
-
-.totalBox {
-  display: flex;
-  justify-content: center;
-  align-content: center;
-  margin: 30px;
-  gap: 20px;
-  font-size: 20px;
-}
-
-.logType {
-  display: flex;
-  font-size: 24px;
-  font-weight: 800;
-}
-
-.totalFocusValue {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: var(--red_F50E0E);
-}
-
-.totalPointValue {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: var(--blue_text_418099);
-}
-
+/* LogPage.module.css */
 .logWrapper {
   border-radius: 20px;
   border: 1px solid var(--gray_DDDDDD);
@@ -119,40 +6,9 @@
   min-height: 600px;
 }
 
-.list {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 20px;
-}
-
-.row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 400px;
-  height: 54px;
-  padding: 10px 16px;
-  border-radius: 20px;
-  border: 1px solid var(--green_E1EDDE);
-  background: var(--white_FFFFFF);
-  font-size: 14px;
-  font-weight: 700;
-}
-
-.rowDate {
-  color: var(--gray_818181);
-}
-
-.rowValue {
-  color: var(--green_text_578246);
-  font-size: 16px;
-}
-
-.noData {
-  margin-top: 100px;
-  color: var(--gray_818181);
-}
+/* 필요 시 미디어 쿼리 유지 */
+@media (max-width: 1024px) {}
+@media (max-width: 768px) {}
 
 @media (max-width: 1024px) {
 }
