+++ src/pages/LogPage/components/LogList/LogList.jsx
@@ -1,7 +1,7 @@
-import styles from '../LogPage.module.css';
-import { formattedTime } from '../../../utils/formattedTime';
-import { formatKST } from "../../../utils/formattedDate";
-import { calculateDailyTotals } from '../../../utils/logCalculator';
+import styles from './LogList.module.css';
+import { formattedTime } from '../../../../utils/formattedTime';
+import { formatKST } from "../../../../utils/formattedDate";
+import { calculateDailyTotals } from '../../../../utils/logCalculator';
 
 const LogList = ({ logType, pointLogs, focusLogs }) => {
